backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.28
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.20.0
redis[hiredis]==5.0.4
celery[redis]==5.3.6
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pyjwt[crypto]==2.8.0
pyotp==2.9.0
hvac==2.1.0
pycryptodome==3.20.0
pydantic[email]==2.6.4
pydantic-settings==2.2.1
email-validator==2.2.0
httpx==0.27.0
python-multipart==0.0.9
loguru==0.7.2
orjson==3.10.1
